Stop pirating `scaleminmax`
`scaleminmax` is owned by ImageCore and the colors it operates
on are owned by ColorTypes. In contrast to both Colors and
ColorVectorspace, for which piracy is their raison d'etre,
this is not really a sanctioned case of piracy.

We could add these methods to ImageCore, but I think they'd be a
little dicey, so best to make them internal.