try to get all tests passing
Ran into a new bug related to getting the screen size from the default Gdk display.
screen_size() works early in the tests but fails later on. It looks like the
GListModel of monitors is being unreferenced somewhere, whereas it should be left
alone if we are respecting the introspection annotations (which it looks like we are).