Gtk4 port (#284)
* begin port to gtk4

* try to get all tests passing

Ran into a new bug related to getting the screen size from the default Gdk display.
screen_size() works early in the tests but fails later on. It looks like the
GListModel of monitors is being unreferenced somewhere, whereas it should be left
alone if we are respecting the introspection annotations (which it looks like we are).

* restore some code that was removed during debugging of refcounting problem

* get tests passing again

* fix contrast gui

* try uncommenting player controls code

seeing a freeze in the tests that needs looking into

* bump Gtk4 compat

* bump GtkObservables compat

* set a minimum size for the canvas

When the canvas is put in a window such that it receives no allocated space, it is
never realized, and this causes everything to freeze. It would be good to solve
the underlying issue, but setting a small minimum size (10 by 10 pixels here) prevents
it from ever happening.

* add a delay to get last commmented out test working

* remove rounded corners in frames using CSS

* set `obey_child` to false in AspectFrame

* update README, tweak popup UI, small cleanups